Replace Header method with SetHeader
The previous version of the Header didn't work as intended. The
method returned a shared_ptr that was initialized to nullptr.
The way I assume it was intended to be used was to use shared_ptr::reset()
to set the new linked list entry to the list. However, this also
clears the custom deallocator, so there's no point in providing
this kind of dummy shared_ptr.

The new method adds a vector of strings to the header list and sets
the header as CURLOPT_HTTPHEADER to the handle. The header is stored
as a member variable as it needs to live until the message has been
sent.

Example usage:
```
auto message = Networking::Requests::Client(query);
message.SetHeader({ "Content-Type: application/json", "Accept: application/json" });
auto response = message.Send();
```

Signed-off-by: Joonas Salonpää <joonas.salonpaa@vaisala.com>